feat: Add singleton instance support
