+++ packages/core/test/utils/http_client_test.dart
@@ -2,6 +2,7 @@ import 'package:analytics/analytics.dart';
 import 'package:analytics/logger.dart';
 import 'package:analytics/state.dart';
 import 'package:analytics/utils/http_client.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart';
 import 'package:mockito/mockito.dart';
@@ -12,6 +13,7 @@ import '../mocks/mocks.mocks.dart';
 void main() {
   group("HTTP Client", () {
     setUp(() {
+      TestWidgetsFlutterBinding.ensureInitialized();
       LogFactory.logger = Mocks.logTarget();
     });
     test("It logs on bad response for get Settings", () async {
@@ -19,7 +21,8 @@ void main() {
       when(mockRequest.send()).thenAnswer(
           (_) => Future.value(StreamedResponse(const Stream.empty(), 300)));
       when(mockRequest.url).thenAnswer((_) => Uri.parse("http://segment.io"));
-      HTTPClient client = HTTPClient(Analytics(
+      HTTPClient client = HTTPClient(
+        Analytics.internal(
           Configuration("123", requestFactory: (_) => mockRequest),
           Mocks.store()));
 
@@ -33,7 +36,7 @@ void main() {
       when(mockRequest.send()).thenAnswer(
           (_) => Future.value(StreamedResponse(const Stream.empty(), 300)));
       when(mockRequest.url).thenAnswer((_) => Uri.parse("http://segment.io"));
-      HTTPClient client = HTTPClient(Analytics(
+      HTTPClient client = HTTPClient(Analytics.internal(
           Configuration("123", requestFactory: (_) => mockRequest),
           Mocks.store()));
